first set of patches trying to cleanup some of the ssl stuff that Nalin Dahyabhai <nalin@redhat.com> suggested
When creating certs, there is no reason to populate the country/state/local
info with bogus data. It's redundant, and unneed, so remote it.

Attempt to add support for adding the proper X509 extensions to
indicate what is or is not a CA cert. On my test machine, this
seems to segfault, so it's commented out at the moment.

Set CA key serial number to be a hash of the subject name. Better
than just the static '0' we were using, but still needs work.